Typescriptize current-user service
