+++ app/services/current-user.ts
@@ -1,0 +1,80 @@
+import { computed } from '@ember/object';
+import PromiseProxyMixin from '@ember/object/promise-proxy-mixin';
+import ObjectProxy from '@ember/object/proxy';
+import Service, { inject as service } from '@ember/service';
+import { Promise as EmberPromise } from 'rsvp';
+
+/**
+ * @module ember-osf-web
+ * @submodule services
+ */
+
+/**
+ * Access information about the currently logged in user
+ *
+ * @class current-user
+ * @extends Ember.Service
+ */
+export default class CurrentUserService extends Service {
+    store = service('store');
+    session = service('session');
+
+    /**
+     * If logged in, return the ID of the current user, else return null.
+     *
+     * @property currentUserId
+     * @type {String|null}
+     */
+    currentUserId: string|null = computed('session.data.authenticated', function(this: CurrentUserService) {
+        const session = this.get('session');
+        if (session.get('isAuthenticated')) {
+            return session.get('data.authenticated.id');
+        } else {
+            return null;
+        }
+    });
+
+    /**
+     * Return an observable promise proxy for the currently logged in user. If no user is logged in, resolves to null.
+     *
+     * TODO: Refactor to use a task, not a proxy object
+     *
+     * @property user
+     * @return Promise proxy object that resolves to a user or null
+     */
+    user: User|null|EmberPromise<User|null> = computed('currentUserId', function(this: CurrentUserService) {
+        const ObjectPromiseProxy = ObjectProxy.extend(PromiseProxyMixin);
+        return ObjectPromiseProxy.create({
+            promise: this.load().catch(() => null),
+        });
+    });
+
+    /**
+     * Fetch information about the currently logged in user.
+     * If no user is logged in, this method returns a rejected promise.
+     *
+     * @method load
+     * @return {RSVP.Promise}
+     */
+    load(this: CurrentUserService) {
+        return new EmberPromise((resolve, reject) => {
+            const currentUserId = this.get('currentUserId');
+            if (currentUserId) {
+                const currentUser = this.get('store').peekRecord('user', currentUserId);
+                if (currentUser) {
+                    resolve(currentUser);
+                } else {
+                    this.get('store').findRecord('user', currentUserId).then(user => resolve(user), reject);
+                }
+            } else {
+                reject();
+            }
+        });
+    }
+}
+
+declare module 'ember' {
+    interface ServiceRegistry {
+        'current-user': CurrentUserService;
+    }
+}
